Changed how the two phase commit works.
This is how the two phase commit is used in zope.

    1. tpc_begin(txn)
    2. commit(txn)
    3. tpc_vote(txn)
    4. tpc_finish(txn)

What zalchemy is doing:

  - commit does a session.flush() which actually executes all sql statements.
  - tpc_finish() does a transaction.commit() in the sqlalchemy transaction
  - tpc_abort() does a transaction.rollback() in the sqlalchemy transaction

If commit fails or another DataManager fails, data is not commited to the
database.

Thanks to Michael Bayer (the author of sqlalchemy) who gave me the hint to
integrate sqlalchemy correctly into the two phase commit.



git-svn-id: svn://svn.zope.org/repos/main@73142 62d5b8a3-27da-0310-9561-8e5933582275