Now that we no longer install Zope 3 (neither from SVN nor from a tarball) there'll be no mkzopeinstance script. The old zc.recipe.zope3instance recipe won't work anymore.
Use the new zc.zope3recipes recipes to define an application and an instance.
Unfortunately, this requires us to dump a lot of ZCML into buildout.cfg (that
can be considered an advantage since it provides us with much more control).

An advantage of the zc.zope3recipes:app recipe is that it creates a
"bin/<name_of_instance_section>" script that works like zopectl.

After running the buildout, start Zope with bin/instance fg. Enjoy.


git-svn-id: svn://svn.zope.org/repos/main@74214 62d5b8a3-27da-0310-9561-8e5933582275