Changed registration of file upload data converter. Register FileUploadDataConverter for IBytes/IFileWidget. This will force to use the data converter only for IFileWidget and not for IBytes at all. IBytes/IWidget uses by default the FieldDataConverter.
Added tests

Note:
The default widget for IBytes is the IFileWidget. 

What does this mean:
If you need to use a text area widget for IBytes, you have to
register a custom widget in the form because the default widget
for IBytes is a file upload widget. You can do this by simply
set ``fields['foobar'].widgetFactory = TextWidget``.

git-svn-id: svn://svn.zope.org/repos/main@76629 62d5b8a3-27da-0310-9561-8e5933582275