Tests 15 (propnullns) and 16 (propget) from the props suite of litmus version 10.5 (http://www.webdav.org/neon/litmus/) expose a bug in Zope propertysheet access via DAV.  If a proppatch command sets a property with a null xmlns, e.g. with a PROPPATCH body like:
<?xml version="1.0" encoding="utf-8" ?>
<propertyupdate xmlns="DAV:">
<set>
<prop>
 <nonamespace xmlns="">randomvalue</nonamespace>
</prop>
</set>
</propertyupdate>

When we set properties in the null namespace, Zope turns
around and creates (or finds) a propertysheet with the
xml_namespace of None and sets the value on it.  The
response to a subsequent PROPFIND for the resource will fail
because the XML generated by dav__propstat included a bogus
namespace declaration (xmlns="None").

Fixed by amending OFS.PropertySheets.dav__propstat.



git-svn-id: svn://svn.zope.org/repos/main@76721 62d5b8a3-27da-0310-9561-8e5933582275