A number of fixes to the egg packaging and writing the first test that optionstorage has :(
Also, changing the key back to "optionstorage", so that we don't have to write
a generation script (schooltool's been using optionstorage for a while). I'm
pretty sure noone's downloaded/used this egg yet because the packaging was so
broken.

Thanks to ignas for trying to write the generation script and noticing the
errors.

Also cleanup some whitespace.


git-svn-id: svn://svn.zope.org/repos/main@77386 62d5b8a3-27da-0310-9561-8e5933582275