Continued development on "megrok.trails".  Destroyed the central Trail registry which the initial version contained, in favor of a factory that produces AbsoluteURL factories, which makes things much simpler and requires no global storage beyond the adapter registry itself. Added the beginnings of a functional test suite.
And, most importantly from the point of view of people using the
prototype, it should now work when you ask for the URL of a view atop
a class to which a Trail exists.  Before, it would only work when the
URL was requested for the object itself, because IAbsoluteURL adapters
were created but not named "absolute_url" adapters.  It turns out that
Zope sometimes uses one, and sometimes the other.  This is quite
mysterious to me.  I will have to read more about named adapters
before I understand why they are used intead of interface-identified
adapters.


git-svn-id: svn://svn.zope.org/repos/main@80873 62d5b8a3-27da-0310-9561-8e5933582275