Okay, I am back working on REST again. ;-) This time I am less ambitious.
I have a special publisher for REST and a special request. The code is 
still experimental, but I discovered already several flaws during 
testing, so it is getting better. I also have a little demo for 
folders that provides a very simple RESTive API.

I also developed a REST client that is somewhat similar to test 
browser. It works very well and can be used in real applications as 
well.

Everything is 100% tested!

Thanks to Bayview Financial for allowing me to develop this code under 
the ZPL and publish it to svn.zope.org.



git-svn-id: svn://svn.zope.org/repos/main@82045 62d5b8a3-27da-0310-9561-8e5933582275