(1) default transitive query factories are no longer weak refs--that was insane of me, afaict :-/.
(2) listeners default to not being persistent weakrefs, but can be. 
Making listeners weakrefs at least was not insane, but a bit too
aggressive perhaps.  

(3) added sugar for the spelling to search on an actual relation. 
``None`` was not terribly readable.  You can now spell ``None`` as
``zc.relation.RELATION``, which does seem to read a bit better for this
purpose. 

(4) rearranged order of subsequent docs in README.  tokens.txt is
excessively complicated, so don't want to send anyone into that next. 
Will consider rewriting tokens.txt with less extreme/involved examples.

(5) swictched to using bootstrap external.


git-svn-id: svn://svn.zope.org/repos/main@85272 62d5b8a3-27da-0310-9561-8e5933582275