Reimplement distributionForPackage; its tests now pass.
As suspected, the old algorithm was faulty because it assumed that a module object's __file__ correctly reflected the filepath to that module; however, namespace packages which extend that module's namespace will clobber __file__ so that is unreliable. The new algorithm is roughly::

 * Given a module_dottedname to find the distribution of ... for every entry in sys.path:
  * Reject any entry that's not an egg.
  * Find all distributions contained within the path.
  * For every distribution contained within the path:
   * Reject any distribution that doesn't have module_dottedname in its contained packages
   * Reject any distribution that DOES have module_dottedname in its namespace_packages (declared in the egg metadata)
 * Hope that only one candidate remains unrejected, and return that candidate as correct.

This is much more robust; thanks to dcrosta for helping with it.
  


git-svn-id: svn://svn.zope.org/repos/main@85492 62d5b8a3-27da-0310-9561-8e5933582275