Allow teardown of the functional test layers.
This is safe to do if your layers don't have custom setup that leave state in
global variables that is not cleared by registered CleanUp handlers.

As a result the test runner no longer needs to spawn subprocesses for
subsequent test layers, and therefore you can use pdb and bin/test --coverage
for tests in those layers.



git-svn-id: svn://svn.zope.org/repos/main@87261 62d5b8a3-27da-0310-9561-8e5933582275