A lot of work. This will totally break rdbexample, but it was broken anyway and now we actually have a test. What has happened:
* a doctest in README.txt

* use SQLAlchemy 0.5beta1; this has some declarative functionality we
  need.

* use newest zope.sqlalchemy and z3c.saconfig which we'll use to set up the
  database. (instead of collective.lead)

* use Grok 0.13, which has the Context and Container classes we need.

A few new directives are introduced.

A lot of work remains to be done, especially in the writing of more
tests. more rdb.key situations need to be explored, a lot of the directives
need to be firmed up (better default behavior, more validators, lots more
tests), but at least this is progress.



git-svn-id: svn://svn.zope.org/repos/main@87734 62d5b8a3-27da-0310-9561-8e5933582275