Monkey patch for LP #257276
This code is taken from the encodings module of Python 2.4.
Note that this code is originally (C) CNRI and it is possibly not compatible
with the ZPL and therefore should not live within svn.zope.org. However this
checkin is blessed by Jim Fulton for now. The fix is no longer required with
Python 2.5 and hopefully fixed in Python 2.4.6 release.



git-svn-id: svn://svn.zope.org/repos/main@89727 62d5b8a3-27da-0310-9561-8e5933582275