We want interfaces provided by the client storage to reflect interfaces provided by the served storage on the storage server.
The old way this was done was to copy all of the interface
declarations from the served storage. This has 2 problems:

1. Not all interfaces copied were actually provided by the client
storage. Just because a remote storage provides an interface, doesn't
mean the client storage can, especially if the method (or attribute)
isn't supported by the ZEO protocol.

2. Older clients could get import errors while trying to import copies
interfaces.

Now, we only declare from a known set of interfaces defined on the client.


git-svn-id: svn://svn.zope.org/repos/main@92434 62d5b8a3-27da-0310-9561-8e5933582275