Switched to using the subprocess module so that the zeo server closes all but the standard file descriptors.  This fixed a really hard to debug test failure. (A server started after a client storage was started had the client storage's cache loc file open, preventing the client storage from being reopened.)
For starting servers, the storage config not defaults to a simple
file-storage configuration.

Use a simple log file configuration for the storage server rather than
trying to somehow copy the useless test log configuration. This makes
it far more likely that there will be a sever log available when
debugging test failures.


git-svn-id: svn://svn.zope.org/repos/main@93120 62d5b8a3-27da-0310-9561-8e5933582275