The pack code was not removing as many objects as it should with GC on.
This happened because the pack code was following references from all 
object states in the object_ref table, when it should have been 
following references only from states that will exist after the pack.

The existence of this bug was obscured by an optimization that 
populated only the necessary portion of the object_ref table.  This 
suggests the optimization was a bad choice.  As it turns out, the
object_ref table could be very useful for debugging, and 
populating it is pretty easy, so now we populate the entire object_ref 
table before packing.



git-svn-id: svn://svn.zope.org/repos/main@94984 62d5b8a3-27da-0310-9561-8e5933582275