improve group testing, now we have 100% coverage in every test except the object.py implementation
Note:
the z3c.pt implementation is buggy and doesn't work right.
I left a test in group.txt which shows that a div tag get skipped with z3c.pt.
We need to fix this in z3c.pt or chameleon and the test should be fine without
any changes.

git-svn-id: svn://svn.zope.org/repos/main@98805 62d5b8a3-27da-0310-9561-8e5933582275