Make a branch to support system Pythons better.
Rationale:

Some development shops using zc.buildout want to be able to take advantage of
OS-provided security updates when possible.  Being able to use the system
Python without installed site-packages is a basic goal.

Relying on the operating system for certain Python packages can also help with
keeping up-to-date with security releases.  Moreover, for some Python packages
this can make installation easier.  lxml/libxml2, for instance, is still
painful to install strictly within buildout.

We want to allow a usage of a system Python in buildout.  We also want to make
it possible to use certain packages from the system Python's site-packages.

This is risky.  One of buildout's advantages is repeatability. (That is, as
defined by Jim in his tutorial, "[i]t should be possible to check-in a buildout
specification and reproduce the same software later by checking out the
specification and rebuilding.")  The more you rely on the system software, the
less buildout can have enough control for a repeatable build.

We will divide our goals into two scenarios.

The first scenario is to use a system Python without its site-packages.

The second scenario is to use a system Python with some of its site-packages.

The second scenario may be too difficult; the current plan is that this will
be scaled back to use a system Python with *all* of its site-packages.



git-svn-id: svn://svn.zope.org/repos/main@101292 62d5b8a3-27da-0310-9561-8e5933582275