Speed optimization for match and bobo_handle in the _Handler class.
The match and bobo_handle properties of the class _Handler were meant to be
executed once, and then cache their return value as an optimization; the code
failed to cache the value (properties are not overridden by
self.__dict__[key]), and thus the routes were computed at every access of the
property.

Now the caching works, storing the bobo_handle and match attributes without
using properties (which has also the advantage that the values are computed at
start-up time instead of computing them at the first execution).


git-svn-id: svn://svn.zope.org/repos/main@106746 62d5b8a3-27da-0310-9561-8e5933582275