Add interfaces which can be used for cache response mutation and intercept plus documentation on how they might be used.
Add (back?) ILastModified concept, plus a default adapter for views which delegates to the view's context.

Move registration from an import side-effect to ZCML. Yes, this means you have to register it yourself if you want to use it without ZCML, but you can just as easily just create a global registry object yourself if your'e not bought into Zope's mechanisms of configuration. 

git-svn-id: svn://svn.zope.org/repos/main@107426 62d5b8a3-27da-0310-9561-8e5933582275