Fixed a (distributed) threading issue.
tpc_finish cleared the storage lock before returning.  This meant that
other clients could commit transactions and send invalidations before
the return value was sent to the client. With the recent threading fix
in ClientStorages, getting the tpc_finish result after other
invalidations are sent leads to deadlock on the client.

Also fixed a bug in handling of undo errors.  There's no oid to
associate with the error and the oid doesn't really matter.

Also added a delay repr.


git-svn-id: svn://svn.zope.org/repos/main@108398 62d5b8a3-27da-0310-9561-8e5933582275