Register the TALES translators using grok.
We now have a working Chameleon engine that still supports ``path``,
``exists`` and ``provider`` without ``z3c.pt``, but only as regular
TALES expressions, i.e. as ``tal:content="exists: varname"`` and not
as Python expressions any more
(``tal:content="exists('varname')"``). As this is neighter supported
by z3c.pt anymore, there might be not much loss.

In the long run we might try to get rid of ``path`` and ``exists``
anyway, as they are expression types one usually needs when no
appropriate Python expression is available or it is too cumbersome to
write down the ``python:`` prefix all the time (which is not necessary
with megrok.chameleon).


git-svn-id: svn://svn.zope.org/repos/main@109549 62d5b8a3-27da-0310-9561-8e5933582275