fixed regression in Unauthorized handling:
In previous Zope versions string exceptions were used and all exceptions
re-raised. This doesn't work with each kind of exception, but we can still do
it with Unauthorized exceptions. This way the special handling for Unauthorized
exceptions works again, HTTPResponse._unauthorized is called again. This is now
done after rendering to make sure we don't break logging and custom views.

git-svn-id: svn://svn.zope.org/repos/main@110974 62d5b8a3-27da-0310-9561-8e5933582275