Make test a little bit easier to read.
The call to invalidate seems to be a bit hackish and ignores that fact that
with ZEO store() may return None. It seems that the call to invalidate simply
wants to queue the invalidation message and leverage the locking but doesn't
care too much about the order - it just needs to satisfy the guard. By using
the current transaction number of the Connection we can always ensure that the
guard will pass.




git-svn-id: svn://svn.zope.org/repos/main@113519 62d5b8a3-27da-0310-9561-8e5933582275