Trying to deal with https://bugs.launchpad.net/zodb/+bug/641481
I'm fairly sure that the problem has to do with the fact that ZODB'
object caches implementation has its own weakref implementation.

The approach in this branch is to basically stop using GC for
persistent objects that are in object caches.  This is a band aid. :(

Once consequence is that there will significant leakage if caches
aren't explicitly cleared when they're no-longer used.  In normal
application code, this shouldn't be a problem, since caches typically
are used for the life of a program.  It could be a problem for test
code that open databases and don't close them.

I'm still chasing memory leaks.  It's possible that these are all due
to tests not closing databases (or clearing caches if they are managed
directly.)


git-svn-id: svn://svn.zope.org/repos/main@116780 62d5b8a3-27da-0310-9561-8e5933582275