Add some test coverage that Missing.MV, used for missing values in Products.ZCatalog, can be used in templates without raising errors.
This was an attempt to resolve LP#649343 but I discovered that the
template which was causing the error was using zope.app.pagetemplate
instead of Products.Five.browser.pagetemplatefile.  Once I switched I
no longer for the MissingValue error so the ticket seems to be invalid
or at least I can no longer reproduce it.  No harm in preserving the
test coverage though.


git-svn-id: svn://svn.zope.org/repos/main@117824 62d5b8a3-27da-0310-9561-8e5933582275