Delete OID blobs only when it's the last reference.
When using undo, we copy the oid 'forward', creating an additional reference to the blob (but not duplicating the data). Only unlink if this is the last reference. We need to make this a BEFORE trigger, otherwise we cannot reliably count on a full DELETE FROM TABLE (cnt would be 0 at all times). This removes the need to catch the exception (caused by multiple refs). Also, clean out oid refs on drop_all().

git-svn-id: svn://svn.zope.org/repos/main@121933 62d5b8a3-27da-0310-9561-8e5933582275