Fixed issue where a template would not have a ``_v_macros`` attribute.
This volatile attribute is relied upon by some legacy code such as the
Zope 2 product ``PageTemplates``.

In this changeset, the interface is changed to match that of the TAL
parser's ``getCode`` method which returns a tuple ``(program,
macros)``. These are in turn assigned to the volatile attributes
``_v_program`` and ``_v_macros``.


git-svn-id: svn://svn.zope.org/repos/main@122788 62d5b8a3-27da-0310-9561-8e5933582275