Initial import of mongopersist code. Here is an except fromt he README:
``mongopersist`` is a Mongo storage implementation for persistent
Python objects. It is *not* a storage for the ZODB.

The goal of ``mongopersist`` is to provide a data manager that serializes
objects to Mongo at transaction boundaries. The mongo data manager is a
persistent data manager, which handles events at transaction boundaries (see
``transaction.interfaces.IDataManager``) as well as events from the
persistency framework (see ``persistent.interfaces.IPersistentDataManager``).



git-svn-id: svn://svn.zope.org/repos/main@123285 62d5b8a3-27da-0310-9561-8e5933582275