While importing toolset.xml, only warn about missing required tools.
We print a warning when the class of a required tool is not found and
continue with the next tool.  The previous behaviour could break the
install or uninstall of any add-on, as the missing class may easily be
from a different unrelated add-on that is no longer available in the
zope instance.

This merges r123869 from trunk.


git-svn-id: svn://svn.zope.org/repos/main@123871 62d5b8a3-27da-0310-9561-8e5933582275