Have to clear cached envs to most faithfully reproduce existing behavior.
For example, when doing a develop egg and subsequently using it,
buildout will still have to "get" the dist and find it as an egg
link.  If the env is magically updated when the egg is developed, then
it no longer has to "get" the dist since it knows it has it.  This
fixed the last of the test failures.

git-svn-id: svn://svn.zope.org/repos/main@124123 62d5b8a3-27da-0310-9561-8e5933582275