Option: use only import / export steps explicitly names by the current profile.
If this option is set, the setup tool ignores any steps which might be
globally registered:  it uses only those named in the profile's
'import_steps.xml' / 'export_steps.xml' files.

This is particularly useful for configuring sites with baseline profiles,
where arbitrary add-on steps are not only useless, but potentially damaging.


git-svn-id: svn://svn.zope.org/repos/main@124484 62d5b8a3-27da-0310-9561-8e5933582275