Fix for windows compatibility (hopefully)
This test failed under windows. 
The code to be tested failed with the following
comparison:
os.path.normpath('/absolute') == os.path.abspath('/absolute')

even when one tries to reproduce the failure unter
linux with ntpath, the failure does not appear.
Under windows, the drive letter must be part
of the absolute path, in this is what abspath
but not normpath does.


git-svn-id: svn://svn.zope.org/repos/main@126448 62d5b8a3-27da-0310-9561-8e5933582275