Export dictionaries per worker rather than a tuple, add in the worker version.
This is backwards incompatible in that it changes the format of the data the
status server sends. Some kind of backwards incompatibility is required to be
able to export the worker version as well. The change to a dictionary is an
attempt to try prevent this happening again in future.


git-svn-id: svn://svn.zope.org/repos/main@127386 62d5b8a3-27da-0310-9561-8e5933582275