Strip white space from name when adding a column or index.
I have seen this in a live site and it is tricky to figure out what is wrong.
Solution if it is too late: delete the column or index and recreate without the space.
For the column, rebuild the catalog.  For the index, index that particular index.
This is merged from trunk, r127898.

git-svn-id: svn://svn.zope.org/repos/main@127899 62d5b8a3-27da-0310-9561-8e5933582275