The ZCML directive handler for a translation domain registration now loads message catalogs lazily.
If the `zope_i18n_allowed_languages` environment variable is not
provided, the allowed languages are determined at runtime. To
participate in this, a translation domain implementation must now
expose a set of languages in a ``languages`` attribute.

The difficulty in this patch is that the ``INegotiator`` interface
relies on an "allowed languages" set, but this set is not readily
available. In the current definition of a translation domain, we can't
know until we attempt a translation whether a language is supported.

The solution in this patch is to require that a translation domain
exposes the defined languages and then query all translation domains
to determine the allowed languages. It would be nice if we could just
provide a value of ``None`` (meaning any language), but there's
already existing implementations that expect a set.


git-svn-id: svn://svn.zope.org/repos/main@127957 62d5b8a3-27da-0310-9561-8e5933582275