FileCache reliability fixes.
Avoid incorrect cache lookups (or invalidations) when a username is a
proper prefix of some other username.

Avoid cache poisoning if usernames contain embedded '::' separators or
newlines.

Avoid exceptions on a race condition if the cache file disappears after
we check for its existence but before we open it for reading.

git-svn-id: svn://svn.zope.org/repos/main@127958 62d5b8a3-27da-0310-9561-8e5933582275